% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_beta.R
\name{add_beta}
\alias{add_beta}
\title{Add a new component to a mixture distribution.}
\usage{
add_beta(
  p,
  mixm,
  opt_method = "L-BFGS-B",
  monotone = TRUE,
  min_null = 0.4,
  suppress_warn = TRUE,
  ...
)
}
\arguments{
\item{p}{A vector of p-values}

\item{mixm}{A \code{betamix} object representing the beta mixture model.
This is typically created by the fit_betamix function.}

\item{opt_method}{Optimization method. Options are "L-BFGS-B" or "SANN".}

\item{monotone}{Only use monotonically-decreasing beta components in the
mixture model}

\item{min_null}{The lower bound on the null distribution mixing fraction.}

\item{...}{Additional parameters to be passed to \code{bbmle::mle2}.}
}
\description{
Adds a monotonically decreasing beta distribution to an
existing distribution.
}
