% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/alt_posterior.R
\name{alt_posterior}
\alias{alt_posterior}
\title{Compute the posterior probability of the alternative hypothesis}
\usage{
alt_posterior(p, mixm)
}
\arguments{
\item{p}{A numeric vector of p-values}

\item{mixm}{A \code{betamix} object representing the beta mixture model}
}
\description{
Maps a vector of p-values to a vector of posterior probabilities
given a beta mixture model.
}
