% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bayes_p.R
\name{bayes_p}
\alias{bayes_p}
\title{Map p-values to posterior probabilities}
\usage{
bayes_p(p, n_boots = 1000, alpha = 0.01, n_cores = 1, subsample = 1,
  level_p = FALSE, max_comp = 5, opt_method = "L-BFGS-B", ...)
}
\arguments{
\item{p}{A numeric vector of p-values'}

\item{n_boots}{A number providing the number of bootstraps used
to calculate the convergence statistic.}

\item{alpha}{A number representing the convergence statistic.}

\item{n_cores}{A number representing the number of cores to use
for the bootstrap calculation.}

\item{subsample}{The proportion of p to subsample to calculate
the beta mixture}

\item{level_p}{Logical. Apply \code{Polyfit::levelPvalues} to \code{p}
before the fitting procedure.}

\item{max_comp}{A number representing the maximum number of non-uniform
components to include in the mixture distribution.}

\item{opt_method}{Optimization method. Options are "L-BFGS-B" or "SANN".}

\item{...}{Additional parameters to be passed to \code{bbmle::mle2}.}
}
\description{
Given a vector of p-values from independent hypothesis
tests, calculate the Bayesian posterior probability of the alternative
hypothesis being true. The method implemeted is approximately that
which is described by Erikson et al., 2006 and Allison et al., 2002.
The only difference being that the mixture of beta distribtutions used
to model the distribution of p-values only permits monotonically decreasing
betas.
}
