% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bootstrap_Q.R
\name{bootstrap_Q}
\alias{bootstrap_Q}
\title{Compute statistics for model selection}
\usage{
bootstrap_Q(p, c0, c1, n_boots, n_cores)
}
\arguments{
\item{p}{A numeric vector of p-values}

\item{c0}{A model with n beta components}

\item{c1}{A model with n+1 beta components}

\item{n_boots}{Number of bootstraps to perform}

\item{n_cores}{Number of cores to use to perform the bootstrapping}
}
\description{
Computes statistics to assess the value of adding an additional beta component
}
