% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compare_betamixes.R
\name{compare_betamixes}
\alias{compare_betamixes}
\title{Compute statistics for model selection}
\usage{
compare_betamixes(p, mixm0, mixm1, n_boots, n_cores)
}
\arguments{
\item{p}{A numeric vector of p-values}

\item{mixm0}{A \code{betamix} object that defines a uniform-beta mixture to be tested against}

\item{mixm1}{A \code{betamix} object that defines a uniform-beta mixture to be tested}

\item{n_boots}{Number of bootstraps to perform}

\item{n_cores}{Number of cores to use to perform the bootstrapping}
}
\description{
Compute the p-value for the improvement of one uniform-beta mixture over another
with respect to a given set of p-values
}
