% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/construct_nll.R
\name{.construct_nll}
\alias{.construct_nll}
\title{Generate a negative log-likelihood function for a uniform-beta mixture model.}
\usage{
.construct_nll(arg_names)
}
\arguments{
\item{arg_names}{A character vector with the names of the parameters
of the beta mixture.}
}
\description{
Given a set of argument names, generate a negative
log-likelihood function to be used in the optimization step of
the \code{add_beta} function.
}
