% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sann_generate.R
\name{.sann_generate}
\alias{.sann_generate}
\title{Generate a new point for simulated annealing}
\usage{
.sann_generate(...)
}
\arguments{
\item{...}{The arguments to the negative log likelihood function}
}
\description{
Generate a new point for simulated annealing
}
