% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estimate_params.R
\name{estimate_params}
\alias{estimate_params}
\title{Fit a mixture of beta distributions to data.}
\usage{
estimate_params(p, n_boots = 1000, alpha = 0.01, n_cores = 1,
  subsample = 1, max_comp = 5, opt_method = "L-BFGS-B", ...)
}
\arguments{
\item{p}{A numeric vector of p-values'}

\item{n_boots}{A number providing the number of bootstraps used
to calculate the convergence statistic.}

\item{alpha}{A number representing the convergence statistic.}

\item{n_cores}{A number representing the number of cores to use
for the bootstrap calculation.}

\item{subsample}{The proportion of p to subsample to calculate
the beta mixture}

\item{max_comp}{A number representing the maximum number of non-uniform
components to include in the mixture distribution.}

\item{opt_method}{Optimization method. Options are "L-BFGS-B" or "SANN".}

\item{...}{Additional parameters to be passed to \code{bbmle::mle2}.}
}
\description{
Estimates the parameters of a mixture of beta distributions,
given some data. The mixture only permits monotonically decreasing
betas and a single uniform distribution.
}
