% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit_beta_mix.R
\name{fit_beta_mix}
\alias{fit_beta_mix}
\title{Fit a uniform-beta mixture to a distribution of p-values.}
\usage{
fit_beta_mix(
  p,
  n_boots = 1000,
  alpha = 0.01,
  n_cores = 1,
  subsample = 1,
  max_comp = 4,
  min_null = 0.4,
  opt_method = "L-BFGS-B",
  monotone = TRUE,
  ...
)
}
\arguments{
\item{p}{A numeric vector of p-values}

\item{n_boots}{A number providing the number of bootstraps used
to calculate the convergence statistic.}

\item{alpha}{A number representing the convergence criterion for formulating the mixture.}

\item{n_cores}{A number representing the number of cores to use
for the bootstrap calculation.}

\item{subsample}{The proportion of p to subsample to calculate the beta mixture.}

\item{max_comp}{A number representing the maximum number of non-uniform
components to include in the mixture distribution.}

\item{min_null}{The lower bound on the null distribution mixing fraction.}

\item{opt_method}{Optimization method. Options are "L-BFGS-B" or "SANN".}

\item{monotone}{Logical. Only use monotonically-decreasing beta components in the
mixture model.}

\item{...}{Additional parameters to be passed to \code{bbmle::mle2}.
export}
}
\description{
Estimates the parameters of a uniform-beta mixture distribution,
given some data. The mixture permits multiple betas and a single uniform distribution.
}
