% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/level_p.R
\name{level_p}
\alias{level_p}
\title{Levels the uniform component of a p-value distribution.}
\usage{
level_p(p)
}
\arguments{
\item{p}{A numeric vector of p-values}
}
\description{
The uniform component of a p-distribution is leveled using a cubic
polynomial. The method is similar to that of \code{Polyfit::levelPValues}.
}
