% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/loglike.R
\name{loglike}
\alias{loglike}
\title{Calculate the value of the log likelihood function}
\usage{
loglike(p, coefs)
}
\arguments{
\item{p}{A numeric vector of p-values}

\item{model}{The beta mixture model represented as a named list of distribution parameters.}
}
\description{
Given a distribution of p-values and mixture model coefficients
calculate the value of the log likelihood function
}
