% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/loglike_ubeta.R
\name{loglike_ubeta}
\alias{loglike_ubeta}
\title{Calculate the value of the log likelihood function.}
\usage{
loglike_ubeta(p, mixm)
}
\arguments{
\item{p}{A numeric vector of p-values}

\item{mixm}{An object of class betamix representing the uniform-beta mixture model.}
}
\description{
Given a distribution of p-values and parameters for a beta-uniform mixture,
calculate the value of the log likelihood function.
}
