% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sample_beta_mix.R
\name{sample_beta_mix}
\alias{sample_beta_mix}
\title{Sample a mixture of beta distributions}
\usage{
sample_beta_mix(n, coefs)
}
\arguments{
\item{n}{Number of observations}

\item{coefs}{A named list providing the parameters of the beta mixture}
}
\description{
Draw a random sample from a beta mixture model
}
