% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sample_beta_mix.R
\name{sample_betamix}
\alias{sample_betamix}
\title{Sample from a uniform-beta mixture distribution}
\usage{
sample_betamix(mixm, n)
}
\arguments{
\item{mixm}{A betamix object defining the beta mixture}

\item{n}{Number of draws}
}
\description{
Random draws from a uniform-beta mixture model
}
