% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_beta.R
\name{add_beta}
\alias{add_beta}
\title{Add a new component to a mixture distribution.}
\usage{
add_beta(p, coefs, sann = FALSE, ...)
}
\arguments{
\item{p}{A numeric vector of p-values}

\item{coefs}{A named numeric vector of coefficients representing the
initial model.}

\item{...}{Additional parameters to be passed to \code{bbmle::mle2}.}

\item{opt_method}{Optimization method (see \code{?optim} for options).}
}
\description{
Adds a monotonically decreasing beta distribution to an
existing distribution.
}
