% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/alt_posterior.R
\name{alt_posterior}
\alias{alt_posterior}
\title{Compute the posterior probability of the alternative hypothesis}
\usage{
alt_posterior(p, model)
}
\arguments{
\item{p}{A numeric vector of p-values}

\item{model}{The beta mixture model represented as a named list of distribution parameters.}
}
\description{
Maps a vector of p-values to a vector of posterior probabilites
given a beta mixture model.
}
